backend/requirements.txt
ansible==9.5.1
ansible-core==2.16.6
ansible-lint==5.4.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
audioread==3.0.1
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
bracex==2.2.1

certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
commonmark==0.9.1
contourpy==1.3.1
cryptography==3.4.8
cupshelpers
cycler==0.12.1
dbus-python==1.2.18
debtcollector==2.3.0
debugpy==1.8.12
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.13
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.1.0
docker==7.1.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.17.1
duplicity==0.8.21
enrich==1.2.6
exceptiongroup==1.2.2
executing==2.2.0
fail2ban==0.11.2
fasteners==0.14.1
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.55.1
future==0.18.2
gpg==1.16.0
greenlet==1.1.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.31.0
iso8601==1.0.2
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
jsonschema==3.2.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jwcrypto==1.0
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lockfile==0.12.2
louis==3.20.0
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
numba==0.61.0
numpy==2.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.11.0.86
oslo.config==8.8.0
oslo.context==4.1.0
oslo.i18n==5.1.0
oslo.log==4.6.1
oslo.serialization==4.2.0
oslo.utils==4.12.2
packaging==21.3
pandas==2.2.3
paramiko==2.9.3
parso==0.8.4
pathspec==0.9.0
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.50
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pykerberos==1.1.14
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
pyrsistent==0.20.0
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
pyspark==3.5.4
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==0.21.1
python-novnc==1.0.0
pytz==2022.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.1
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
reportlab==3.6.8
requests==2.32.3
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
resolvelib==0.8.1
rfc3986==1.5.0
rich==11.2.0
roman==3.3
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.1
shiboken6==6.7.0
simpleaudio==1.0.4
simplejson==3.17.6
six==1.16.0
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
SQLAlchemy==1.4.31
sqlparse==0.5.3
stack-data==0.6.3
stevedore==3.5.0
systemd-python==234
tenacity==6.3.1
termcolor==2.3.0
texttable==1.7.0
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
wcmatch==8.3
wcwidth==0.2.13
websocket-client==0.59.0
websockify==0.10.0
Werkzeug==3.1.3
WoeUSB-ng==0.2.12
wrapt==1.13.3
wxPython==4.0.7
xdg==5
xkit==0.0.0
xlrd==2.0.1
xmltodict==0.12.0
yamllint==1.26.3
zipp==1.0.0
